Add test script and number ordinals from 0

+++ benchmarks/impls/c_template/hasher.c
@@ -1,3 +1,6 @@
-uint64_t lookup (const char *str, size_t len) {
+#include <stddef.h>
+#include <stdint.h>
+
+uint32_t lookup (const char *str, size_t len) {
     return len + str[0];
 }
